requirements.txt
faker
pandas
groq
python-dotenv
autogen-agentchat~=0.2
streamlit
